Move widgets to their own namespace
